Improved project modularity
