+++ .gitignore
@@ -1,3 +1,10 @@
+# Editor artifacts
+
+.vscode
+
+
+
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
